Changes to implement suggestions by D. Lange
