+++ EventFilter/DTRawToDigi/python/dturosunpackerSliceTest_cfi.py
@@ -1,6 +1,6 @@
 import FWCore.ParameterSet.Config as cms
 
-dturosunpacker2 = cms.EDProducer("DTuROSRawToDigi2",
+dturosunpacker2 = cms.EDProducer("DTuROSRawToDigiSliceTest",
                                    DTuROS_FED_Source = cms.InputTag("source"),
                                    feds     = cms.untracked.vint32( 1368,),
                                    debug = cms.untracked.bool(True),
