--- EventFilter/DTuROSRawToDigi/plugins/BuildFile.xml
@@ -1,17 +1,0 @@
-<use name="FWCore/Framework"/>
-<use name="FWCore/Utilities"/>
-<use name="FWCore/ParameterSet"/>
-<use name="FWCore/PluginManager"/>
-<use name="FWCore/MessageLogger"/>
-<use name="DataFormats/FEDRawData"/>
-<use name="DataFormats/HcalDigi"/>
-<use name="DataFormats/EcalDigi"/>
-<use name="DataFormats/L1CaloTrigger"/>
-<use name="DataFormats/L1DTTrackFinder"/>
-<use name="DataFormats/L1GlobalMuonTrigger"/>
-<use name="DataFormats/MuonDetId"/>
-<use name="DataFormats/DTDigi"/>
-<use name="CondFormats/DTObjects"/>
-<use name="CondFormats/DataRecord"/>
-<use name="boost"/>
-<flags EDM_PLUGIN="1"/>
